+++ src/App.js
@@ -2,10 +2,10 @@ import React, { useState } from 'react';
 import './App.css';
 import GlobalStyle from './styles/globalStyle';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { Home } from './components/home/Home';
-import { Sign } from './components/sign/Sign';
+import { Home } from './pages/home/Home';
+import { Sign } from './pages/sign/Sign';
 import UserContext from './contexts/UserContext';
-import { Store } from './components/store/Store';
+import { Store } from './pages/store/Store';
 
 function App() {
