+++ src/components/sidebar/Sidebar.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+
+export function Sidebar () {
+ 
+    return(
+        <div>
+            
+        </div>
+    )
+}
